feat: :pencil2: new modifications

+++ Dia2/Ejercicio5_Dia2_MonsalveJohan.py
@@ -3,6 +3,8 @@
 # ####################
 
 print ("Bienvenido a nuestro programa para pasar de celsius a fahrenheit ")
-celsius=  int(input("Dime los grados Celsius: "))
+celsius=  float(input("Dime los grados Celsius: "))
 fahrenheit = (celsius * 9/5) + 32
 print(" la temparatura en Fahrenheit es: " + str (fahrenheit))
+
+# Desarrollado por : Johan Monsalve C.C 1097911956
